trt mode

+++ videoanalyst/model/common_opr/common_block.py
@@ -84,12 +84,13 @@ def xcorr_depthwise(x, kernel):
     torch.Tensor
         cross-correlation result
     """
-    batch = kernel.size(0)
-    channel = kernel.size(1)
-    x = x.view(1, batch * channel, x.size(2), x.size(3))
-    kernel = kernel.view(batch * channel, 1, kernel.size(2), kernel.size(3))
+    batch = int(kernel.size(0))
+    channel = int(kernel.size(1))
+    x = x.view(1, int(batch * channel), int(x.size(2)), int(x.size(3)))
+    kernel = kernel.view(batch * channel, 1, int(kernel.size(2)),
+                         int(kernel.size(3)))
     out = F.conv2d(x, kernel, groups=batch * channel)
-    out = out.view(batch, channel, out.size(2), out.size(3))
+    out = out.view(batch, channel, int(out.size(2)), int(out.size(3)))
     return out
